+++ packages/client/src/pages/Home/components/index.ts
@@ -1,0 +1,3 @@
+export { default as LogoutPageMenu } from './LogoutPageMenu'
+export { default as MainPageMenu } from './MainPageMenu'
+export { default as YandexOAuthButton } from './YandexOAuthButton'
